streamlit-langchain/requirements.txt
# pip3 install -r requirements.txt
streamlit
openai
langchain
tiktoken
cassandra-driver
cassio